types refactor, remove dangling connections
