+++ types/db/Address.ts
@@ -1,0 +1,7 @@
+export default interface Address {
+  id?: number
+  street_address: string
+  admin_level_2: string
+  postal_code: string
+  country: string
+}
