+++ types/db/ParticipantDB.ts
@@ -1,0 +1,9 @@
+export default interface ParticipantDB {
+  id?: number
+  first_name: string
+  last_name: string
+  email: string
+  phone: string
+  address: number
+  team: number
+}
