+++ types/db/RobotDB.ts
@@ -1,0 +1,7 @@
+export default interface RobotDB {
+  id?: number
+  robot_no?: number
+  name: string
+  team: number
+  competition: string
+}
