+++ types/db/Team.ts
@@ -1,0 +1,4 @@
+export default interface Team {
+  id?: number
+  name: string
+}
